Collapse app/platform/ folder; move haptics into systems/ (refs #1200)
Per the canonical 5-folder allowlist directive (2026-05-15T06:24:15Z), the
'platform' subfolder is no longer sanctioned: anything under app/ must be one
of components/, entities/, systems/, tags/, or util/. The 08:38:49Z drift
comment on #1200 calls out app/platform/ as the next folder to collapse,
following the same minimal-move pattern used for input/ (#1216), audio/
(#1217), and session/ (#1218).

Moves (history preserved via git mv):

  app/platform/haptics_backend.{h,cpp}         -> app/systems/haptics_backend.{h,cpp}
  app/platform/ios/haptics_ios_bridge.h        -> app/systems/haptics_ios_bridge.h
  app/platform/ios/haptics_ios_bridge_ios.mm   -> app/systems/haptics_ios_bridge_ios.mm
  app/platform/ios/haptics_ios_bridge_stub.cpp -> app/systems/haptics_ios_bridge_stub.cpp

Updates:
- Relative include paths inside the moved files adjusted for the new depth.
- Caller includes updated (game_loop.cpp, systems/haptic_system.cpp,
  ui/screen_controllers/settings_screen_controller.cpp, tests/test_haptics_backend.cpp).
- CMakeLists.txt: drop the PLATFORM_SOURCES glob; the relocated .cpp files
  are now picked up by the existing SYSTEM_SOURCES glob. Update the iOS
  filter + path in both the iOS bridge compile-check target and the
  SHAPESHIFTER_IOS branch.
- Empty app/platform/ and app/platform/ios/ directories removed.

The platform::haptics and platform::ios namespaces are preserved verbatim,
matching the minimal-move precedent established by #1216–#1218.

Validation:
- Unity Clang Release build: green, zero warnings (-Wall -Wextra -Werror).
- Non-unity Clang Release build: green, zero warnings.
- ctest: 912 passed / 1 skipped (default-skipped smoke) / 0 failed.
- iOS bridge ObjC++ compile gate (SHAPESHIFTER_IOS_BRIDGE_COMPILE_CHECK=ON,
  iphonesimulator SDK): green; CI's 'grep haptics_ios_bridge_ios.mm
  build-ios-bridge' check still finds the bridge in the iOS compile graph.
